+++ unfinished/cocalc/cocalc_test.py
@@ -1,0 +1,8 @@
+import cocalc
+
+
+response = cocalc.Completion.create(
+    prompt = 'hello world'
+)
+
+print(response)
